--- AbMath/Calculator/Simplifications/LogSimplifications.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace AbMath.Calculator.Simplifications
-{
-    class LogSimplifications
-    {
-    }
-}
